First refactor - moved classes to relevant sub-packages (activities, fragments, etc) Added blank Application class
