+++ .idea/modules.xml
@@ -3,6 +3,8 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/itMeet.iml" filepath="$PROJECT_DIR$/itMeet.iml" />
       <module fileurl="file://$PROJECT_DIR$/itMeet.iml" filepath="$PROJECT_DIR$/itMeet.iml" />
     </modules>
   </component>
