+++ app/src/main/java/com/alfagroup/itmeetapp/Activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.alfagroup.itmeetapp;
+package com.alfagroup.itmeetapp.Activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -7,6 +7,10 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import com.alfagroup.itmeetapp.Logic.DbHelper;
+import com.alfagroup.itmeetapp.R;
+import com.alfagroup.itmeetapp.Logic.Session;
 
 public class LoginActivity extends AppCompatActivity implements View.OnClickListener{
     private Button login;
