+++ app/src/main/java/com/alfagroup/itmeetapp/Activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.alfagroup.itmeetapp;
+package com.alfagroup.itmeetapp.Activities;
 
 import android.content.Intent;
 import android.support.design.widget.TabLayout;
@@ -8,6 +8,10 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.alfagroup.itmeetapp.R;
+import com.alfagroup.itmeetapp.Logic.Session;
+import com.alfagroup.itmeetapp.Logic.SimpleFragmentPagerAdapter;
 
 public class MainActivity extends AppCompatActivity {
     private Button btnLogout;
