+++ app/src/main/java/com/alfagroup/itmeetapp/Activities/RegisterActivity.java
@@ -1,4 +1,4 @@
-package com.alfagroup.itmeetapp;
+package com.alfagroup.itmeetapp.Activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -8,6 +8,10 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import com.alfagroup.itmeetapp.Logic.DbHelper;
+import com.alfagroup.itmeetapp.R;
+import com.alfagroup.itmeetapp.Logic.Session;
 
 
 public class RegisterActivity extends AppCompatActivity implements View.OnClickListener{
