+++ app/src/main/java/com/alfagroup/itmeetapp/Fragments/ChosenListFragment.java
@@ -1,28 +1,34 @@
-package com.alfagroup.itmeetapp;
+package com.alfagroup.itmeetapp.Fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageButton;
 import android.widget.ListView;
 import android.app.Activity;
 
 
+import com.alfagroup.itmeetapp.Logic.DbHelper;
+import com.alfagroup.itmeetapp.Logic.MyChosenListViewAdapter;
+import com.alfagroup.itmeetapp.R;
+import com.alfagroup.itmeetapp.Logic.TimeAndDate;
 
 import java.util.ArrayList;
 
 
-public class PersonMatchFragment extends Fragment {
+public class ChosenListFragment extends Fragment {
 
     private DbHelper db;
     public static Activity myActivity;
 
-    public PersonMatchFragment() {
+    public ChosenListFragment() {
     }
-    MyPersonMatchRecyclerViewAdapter adapter;
+    MyChosenListViewAdapter adapter;
     ListView listView;
-    ArrayList<PersonMatch> personMatchArrayList;
+    ImageButton imageButton;
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -33,21 +39,22 @@ public class PersonMatchFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        final View view = inflater.inflate(R.layout.fragment_personmatch_list, container, false);
+        final View view = inflater.inflate(R.layout.fragment_chosen_list, container, false);
         db = DbHelper.getInstance(view.getContext());
 
-        refreshList(view);
+      refreshList(view);
 
 
         return view;
     }
 
+
     public void refreshList(View view){
-        ArrayList<PersonMatch> personMatchArrayList =db.getPersonMatch(getContext());
-        adapter = new MyPersonMatchRecyclerViewAdapter(getContext(), personMatchArrayList);
-        listView = (ListView) view.findViewById(R.id.listPersonMatch);
+        ArrayList<TimeAndDate> timeAndDateArrayList =db.getListDateAndTime(getContext());
+        adapter = new MyChosenListViewAdapter(getContext(), timeAndDateArrayList);
+        listView = (ListView) view.findViewById(R.id.listDateAndTime);
         listView.setAdapter(adapter);
-        adapter.notifyDataSetChanged();
+
 
     }
