+++ app/src/main/java/com/alfagroup/itmeetapp/ItMeet.java
@@ -1,0 +1,17 @@
+package com.alfagroup.itmeetapp;
+
+import android.app.Application;
+
+/**
+ * Created by ilanb on 26/8/2017.
+ */
+
+public class ItMeet extends Application {
+	@Override
+	public void onCreate()
+	{
+		super.onCreate();
+
+
+	}
+}
