+++ app/src/main/java/com/alfagroup/itmeetapp/Logic/Session.java
@@ -1,4 +1,4 @@
-package com.alfagroup.itmeetapp;
+package com.alfagroup.itmeetapp.Logic;
 
 import android.content.Context;
 import android.content.SharedPreferences;
