+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
-    tools:context="com.alfagroup.itmeetapp.LoginActivity">
+    tools:context="com.alfagroup.itmeetapp.Activities.LoginActivity">
 
     <TextView
         android:layout_width="wrap_content"
