+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
-    tools:context="com.alfagroup.itmeetapp.MainActivity">
+    tools:context="com.alfagroup.itmeetapp.Activities.MainActivity">
 
     <android.support.design.widget.TabLayout
         android:id="@+id/tabs"
