+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
-    tools:context="com.alfagroup.itmeetapp.RegisterActivity">
+    tools:context="com.alfagroup.itmeetapp.Activities.RegisterActivity">
 
     <TextView
         android:layout_width="wrap_content"
