+++ app/src/main/res/layout/fragment_personmatch_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="2dp"
     android:layout_marginRight="2dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context="com.alfagroup.itmeetapp.PersonMatchFragment"
+    tools:context="com.alfagroup.itmeetapp.Fragments.PersonMatchFragment"
     tools:listitem="@layout/fragment_personmatch" />
